Remove cyclic imports of schemaBuilder.ts
