+++ experimental/dds/tree2/src/feature-libraries/typed-schema/internal.ts
@@ -4,7 +4,6 @@
  */
 
 // Used by public types, but not part of the desired API surface
-export { RecursiveTreeSchemaSpecification, RecursiveTreeSchema } from "./schemaBuilder";
 
 export { ObjectToMap, WithDefault, Unbrand, UnbrandList, ArrayToUnion } from "./typeUtils";
 
@@ -18,6 +17,8 @@ export {
 	MapSchemaSpecification,
 	LeafSchemaSpecification,
 	MapFieldSchema,
+	RecursiveTreeSchemaSpecification,
+	RecursiveTreeSchema,
 } from "./typedTreeSchema";
 
 export {
