+++ experimental/dds/tree2/src/feature-libraries/typed-schema/typedTreeSchema.ts
@@ -4,7 +4,14 @@
  */
 
 import { assert } from "@fluidframework/core-utils";
-import { EmptyKey, FieldKey, TreeSchemaIdentifier, TreeTypeSet, ValueSchema } from "../../core";
+import {
+	Adapters,
+	EmptyKey,
+	FieldKey,
+	TreeSchemaIdentifier,
+	TreeTypeSet,
+	ValueSchema,
+} from "../../core";
 import {
 	MakeNominal,
 	Assume,
@@ -12,11 +19,12 @@ import {
 	_InlineTrick,
 	FlattenKeys,
 	Named,
+	requireAssignableTo,
 } from "../../util";
 import { FieldKindTypes, FieldKinds } from "../default-field-kinds";
+import { FullSchemaPolicy } from "../modular-schema";
 import { LazyItem, normalizeFlexList } from "./flexList";
 import { ObjectToMap, WithDefault, objectToMapTyped } from "./typeUtils";
-import { RecursiveTreeSchemaSpecification } from "./schemaBuilder";
 
 // TODO: tests for this file
 
@@ -40,6 +48,34 @@ export type NormalizeStructFieldsInner<T extends Fields> = {
 export type NormalizeStructFields<T extends Fields | undefined> = NormalizeStructFieldsInner<
 	WithDefault<T, Record<string, never>>
 >;
+
+/**
+ * Placeholder for to `TreeSchema` to use in constraints where `TreeSchema` is desired but using it causes
+ * recursive types to fail to compile due to TypeScript limitations.
+ *
+ * Using `TreeSchema` instead in some key "extends" clauses cause recursive types to error with:
+ * "'theSchema' implicitly has type 'any' because it does not have a type annotation and is referenced directly or indirectly in its own initializer."
+ *
+ * TODO: how much more specific of a type can be provided without triggering the above error?
+ * @alpha
+ */
+export type RecursiveTreeSchema = unknown;
+
+/**
+ * Placeholder for to `TreeSchemaSpecification` to use in constraints where `TreeSchemaSpecification` is desired but using it causes
+ * recursive types to fail to compile due to TypeScript limitations.
+ *
+ * See `RecursiveTreeSchema`.
+ *
+ * TODO: how much more specific of a type can be provided without triggering the above error?
+ * @alpha
+ */
+export type RecursiveTreeSchemaSpecification = unknown;
+
+{
+	type _check1 = requireAssignableTo<TreeSchemaSpecification, RecursiveTreeSchemaSpecification>;
+	type _check2 = requireAssignableTo<TreeSchema, RecursiveTreeSchema>;
+}
 
 /**
  * T must extend TreeSchemaSpecification.
@@ -294,4 +330,23 @@ export function allowedTypesToTypeSet(t: AllowedTypes): TreeTypeSet {
 	const list: readonly (() => TreeSchema)[] = normalizeFlexList(t);
 	const names = list.map((f) => f().name);
 	return new Set(names);
+}
+
+/**
+ * Schema data that can be be used to view a document.
+ * Strongly typed over its rootFieldSchema.
+ *
+ * @remarks
+ * This type is mainly used as a type constraint to mean that the code working with it requires strongly typed schema.
+ * The actual type used will include detailed schema information for all the types in the collection.
+ * This pattern is used to implement SchemaAware APIs.
+ *
+ * @alpha
+ */
+
+export interface TypedSchemaCollection<T extends FieldSchema = FieldSchema> {
+	readonly rootFieldSchema: T;
+	readonly treeSchema: ReadonlyMap<TreeSchemaIdentifier, TreeSchema>;
+	readonly policy: FullSchemaPolicy;
+	readonly adapters: Adapters;
 }
